All things for sprint 3
5.0: HelpScreen
5.1: Background on Help Screen
5.2: Text explanation of how to play the game
5.3: Scrollable

Additionally, add a back arrow to the help screen class which arrows to go back to the card demo screen

The problem is that I just added the help screen class from the menu screen class and I could not add this to the card demo screen because I don't have that code in my branch. Please help me and change it to the card demo screen when merging our branch.

By the way, the option screen class for sprint 2 was optimized.